requirements.txt
# Unified AI Gateway — Python API (2026 stack)
# Python 3.11+

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
httpx[http2]>=0.28.0
PyJWT[crypto]>=2.9.0
bcrypt>=4.2.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-multipart>=0.0.12
prometheus-client>=0.21.0
cloudinary>=1.41.0
websockets>=14.0
